+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Changelog for NeoFS Node
 ### Changed
 - Optimized GET for fully buffered payloads (#4058)
 - Proxied GET data is not checked against SHA256 hash from header (#4081)
+- SN requests EC part ranges using GET API instead of deprecated RANGE one (#4067)
 
 ### Removed
 - Compression support from FSTree (#4054)